all card info in js
This file contains a single object (maybe not the best idea but...) with every base+ce+prelude card along with machine readable/executable code for all features.

This is a combination of the cards.js plus most of the code from @bafolts tm source. It is HIGHLY edited to make it more useful to me.